Use TRUENAME to fix case discrepancies in translation paths.
On Windows, pathnames are often compared case-sensitively, but shells will
allow you to change to a directory with the wrong case because the
system is still case-insensitive. This causes ASDF to ignore the output
translation that was set because the pathnames don't match.

Note that in order to use TRUENAME on a directory set with --output-dir,
we need to create it first.