Add support for a 'clean' operation.
The idea of the clean operation is to remove the output of other build
operations. This version removes all the dependencies of performing OP on
COMPONENT, as well as the output of performing OP itself. This means that
clean will /not/ function as an "undo" of an operation, but as a full clean
of everything possibly produced by performing an operation.